+++ ui/shared/components/form/LoadOptionsSelect.jsx
@@ -4,16 +4,18 @@ import PropTypes from 'prop-types'
 import { Select } from 'shared/components/form/Inputs'
 import StateDataLoader from '../StateDataLoader'
 
-const parseResponse = optionsResponseKey => responseJson => (
-  { options: responseJson[optionsResponseKey]?.map(value => ({ name: value, value })) }
+const defaultFormatOption = value => ({ name: value, value })
+
+const parseResponse = (optionsResponseKey, formatOption) => responseJson => (
+  { options: responseJson[optionsResponseKey]?.map(formatOption || defaultFormatOption) }
 )
 
 const validateResponse = ({ options }) => (!!options && options.length > 0)
 
-const LoadOptionsSelect = ({ optionsResponseKey, ...props }) => (
+const LoadOptionsSelect = ({ optionsResponseKey, formatOption, ...props }) => (
   <StateDataLoader
     childComponent={Select}
-    parseResponse={parseResponse(optionsResponseKey)}
+    parseResponse={parseResponse(optionsResponseKey, formatOption)}
     validateResponse={validateResponse}
     {...props}
   />
@@ -22,6 +24,7 @@ const LoadOptionsSelect = ({ optionsResponseKey, ...props }) => (
 LoadOptionsSelect.propTypes = {
   url: PropTypes.string.isRequired,
   optionsResponseKey: PropTypes.string.isRequired,
+  formatOption: PropTypes.func,
   errorHeader: PropTypes.string,
   validationErrorHeader: PropTypes.string,
   validationErrorMessage: PropTypes.string,