+++ ui/shared/components/page/PageHeader.jsx
@@ -78,7 +78,7 @@ export default () => (
     {NO_HEADER_PAGES.map(noHeaderRoute)}
     {SIMPLE_HEADER_PAGES.map(simpleHeaderRoute)}
     <Route path="/project/:projectGuid/saved_variants/:variantPage?/:breadcrumbId?/:tag?" component={ProjectSavedVariantsPageHeader} />
-    <Route path="/project/:projectGuid/:breadcrumb/:breadcrumbId?/:breadcrumbIdSection*" component={ProjectPageHeader} />
+    <Route path="/project/:projectGuid/:breadcrumb/:breadcrumbId?/:breadcrumbIdSection?/:breadcrumbIdSubsection*" component={ProjectPageHeader} />
     <Route path="/summary_data/:subPage?" component={SummaryDataPageHeader} />
     <Route path="/variant_search/:pageType/:entityGuid" component={VariantSearchPageHeader} />
     <Route path="/:entity/:entityGuid?/:breadcrumb?/:breadcrumbId*" component={DefaultPageHeaderLayout} />